fixed bugs, moved userserializer to workaround djoser bug
